Changes layout of contact page.
