+++ src/Pages/Contact/Contact.css
@@ -6,5 +6,10 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    margin-top: 10vh;
+}
+
+section {
+  padding-top: 5vh;
+  padding-bottom: 5vh;
+  position: relative;
 }
