--- src/Pages/Contact/Marquee.css
@@ -1,28 +1,0 @@
-.marquee-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    align-content: center;
-    gap: 5rem;
-    width: 100%;
-    height: var(--marquee-height);
-    overflow: hidden;
-    background: var(--middle);
-}
-
-.scrolling-text {
-    animation: marquee 30s linear infinite;
-    font-size: var(--fs-huge);
-    text-align: center;
-    font-family: 'SuisseIntlMono-Bold';
-}
-
-@keyframes marquee {
-    0%   { 
-        transform: translateX(0); 
-    }
-    100% { 
-        transform: translateX(calc(-100% - 5%)); 
-    }
-}
