--- src/Pages/Contact/Marquee.jsx
@@ -1,25 +1,0 @@
-import "./Marquee.css"
-
-
-export default function Marquee() {
-  return (
-    <div className="marquee-container">
-      <a href="mailto:salemsalleh@gmail.com">
-        <p className="scrolling-text">
-          SALEMSALLEH@GMAIL.COM
-        </p>
-      </a>
-      <a href="mailto:salemsalleh@gmail.com">
-        <p className="scrolling-text">
-          SALEMSALLEH@GMAIL.COM
-        </p>
-      </a>
-      <a href="mailto:salemsalleh@gmail.com">
-        <p className="scrolling-text">
-          SALEMSALLEH@GMAIL.COM
-        </p>
-      </a>
-    </div>
-  )
-}
-
