+++ src/Pages/Contact/Social.jsx
@@ -1,0 +1,49 @@
+import './Social.css'
+import React from 'react'
+import { motion } from 'framer-motion'
+import { SocialData } from './SocialData'
+
+const hoverState = {
+    hovered: {
+        scale: 0.95,
+        background: 'var(--accent-2)',
+    },
+    initial: {
+        scale: 1,
+        background: 'var(--darker)',
+    },
+    transition: {
+        duration: 0.3,
+        type: 'tween',
+        ease: 'easeInOut'
+    }
+};
+
+export default function Socialbar() {
+    return (
+        <div className="social">
+            {SocialData.map(({ Name, Icon, Id, Link }) => {
+                return (
+                    <motion.div
+                        className="item-wrapper"
+                        variants={hoverState}
+                        initial="initial"
+                        whileHover="hovered"
+                        whileTap="hovered"
+
+                    >
+                        <h2>{Name}</h2>
+                        <a
+                            key={Id}
+                            href={Link}
+                            className="social-icon"
+                            rel="noreferrer"
+                        >
+                            {Icon}
+                        </a>
+                    </motion.div>
+                )
+            })}
+        </div>
+    )
+}
