--- src/SocialBar/SocialBar.css
@@ -1,16 +1,0 @@
-.social-bar {
-    display: flex;
-    flex-direction: row;
-    position: fixed; 
-    gap: 2rem;
-    margin-left: 2rem;
-    align-items: baseline;
-    font-size: var(--fs-small);
-    margin-right: 5vw;
-  }
-  
-.social-icon {
-    color: var(--light);
-    background-color: transparent;
-    font-size: 1.2rem;
-  }
